[SITIONIX-76] - remove JacksonConfig file and use primary ObjectMapper bean
